+++ src/Kathu/IO/Graphics.hs
@@ -13,12 +13,15 @@ import Data.Functor.Compose
 import qualified Data.Map as Map
 import Data.Text (Text)
 import qualified Data.Text as T
+import qualified Data.Vector as Vec
 import Foreign.C.Types (CInt)
 import GHC.Generics
 import Kathu.Graphics.Color
 import Kathu.Graphics.Drawable
+import Kathu.Graphics.Palette
 import Kathu.IO.Misc
 import Kathu.IO.Parsing
+import Kathu.Util ((>>>=))
 import qualified SDL
 import qualified SDL.Image as SDLI
 
@@ -27,20 +30,23 @@ import qualified SDL.Image as SDLI
 zeroPoint :: SDL.Point SDL.V2 CInt
 zeroPoint = SDL.P $ SDL.V2 0 0
 
-insertImage :: Text -> Image -> SystemLink Image
-insertImage key img = modify (over images $ Map.insert key img) $> img
+instance FromJSON (SystemLink ImageID) where
+    parseJSON (String s) = pure $ (ImageID . fromIntegral) <$> ((flip $ lookupOrAddSL category) adder =<< url)
+        where category = "ImageID"
+              url = fmap T.pack . parseUrl . T.unpack $ s
+              adder = do
+                  url' <- url
+                  modify $ over countingIDs (Map.adjust (mapInsertIncr url') category)
+                  image <- liftSL . SDLI.load . T.unpack $ url'
+                  modify $ over plImages ((flip Vec.snoc) image)
+    parseJSON v          = typeMismatch "ImageID" v
 
--- loads an image if it hasn't already been loaded and add it to our map, otherwise get that
-loadImage :: Text -> SystemLink Image
-loadImage t = do
-    url    <- T.pack <$> (parseUrl . T.unpack $ t)
-    images <- gets (view images)
-    case Map.lookup url images of
-        Just img -> liftSL . pure $ img
-        Nothing  -> (liftSL . SDLI.load . T.unpack $ url) >>= insertImage url
-
-getSurfaceBounds :: MonadIO m => SDL.Surface -> m (SDL.Rectangle CInt)
-getSurfaceBounds s = SDL.surfaceDimensions s >>= pure . SDL.Rectangle zeroPoint
+getSurfaceBounds :: ImageID -> SystemLink (SDL.Rectangle CInt)
+getSurfaceBounds (ImageID iid) = SDL.Rectangle zeroPoint <$> surfaceDim
+    where surfaceDim :: SystemLink (SDL.V2 CInt)
+          surfaceDim = image >>= liftSL . SDL.surfaceDimensions
+          image :: SystemLink Image
+          image = (Vec.!iid) <$> gets (view plImages)
 
 getTextureBounds :: MonadIO m => SDL.Texture -> m (SDL.Rectangle CInt)
 getTextureBounds t = SDL.queryTexture t >>= \t -> pure . SDL.Rectangle zeroPoint $ SDL.V2 (SDL.textureWidth t) (SDL.textureHeight t)
@@ -49,17 +55,16 @@ instance FromJSON AnimationStrip where
     parseJSON = withObject "AnimationStrip" $ \v -> AnimationStrip <$> v .: "id" <*> v .: "frames" <*> v .: "row" <*> v .: "delay"
 
 instance FromJSON (SystemLink Animation) where
-    parseJSON (Object v) = do
-        atlas  <- loadImage <$> v .: "atlas"
-        strips <- v .: "strips"
-        bounds <- (fmap fromIntegral <$> (v .: "bounds" :: Parser (SDL.V2 Int))) :: Parser (SDL.V2 CInt)
-        pure (atlas >>= \a -> pure $ Animation a strips bounds)
+    parseJSON (Object v) = getCompose $ Animation
+        <$> v .:~ "atlas"
+        <*> v .:^ "strips"
+        <*> v .:^ "bounds"
     parseJSON v = typeMismatch "Animation" v
 
 instance FromJSON (SystemLink RenderSprite) where
-    parseJSON (String s)     = pure $ loadImage s >>= \i -> liftSL (getSurfaceBounds i) >>= \bnds -> pure . RSStatic $ StaticSprite i bnds
-    parseJSON obj@(Object _) = (parseJSON obj :: Parser (SystemLink Animation)) >>= \img -> pure $ img >>= \a -> pure . RSAnimated $ AnimatedSprite a 0 0 0
-    parseJSON v              = typeMismatch "RenderSprite" v
+    parseJSON s@(String _) = parseJSON s >>>= \iid -> getSurfaceBounds iid >>= \bnd -> pure . RSStatic . (flip StaticSprite) bnd $ iid
+    parseJSON o@(Object _) = parseJSON o >>>= \iid -> pure . RSAnimated $ AnimatedSprite iid 0 0 0 
+    parseJSON v            = typeMismatch "RenderSprite" v
 
 -- Color
 
@@ -68,3 +73,6 @@ instance ToJSON Color where
 instance FromJSON Color where
     parseJSON (String s) = pure . read . T.unpack $ s
     parseJSON e = typeMismatch "SpecialEntity" e
+
+instance FromJSON Palette where
+    parseJSON = withObject "Palette" $ \v -> Palette <$> v .: "background" <*> pure id