+++ src/Kathu/IO/Parsing.hs
@@ -17,8 +17,11 @@ import Data.Functor.Compose
 import qualified Data.HashMap.Strict as Hash
 import Data.Map (Map)
 import qualified Data.Map as Map
+import Data.Maybe
 import Data.Text (Text)
 import Data.Scientific
+import Data.Vector (Vector)
+import qualified Data.Vector as Vec
 import Kathu.Entity.Damage (DamageProfile)
 import Kathu.Entity.Item (Item)
 import Kathu.Graphics.Drawable
@@ -37,7 +40,7 @@ fromScientific = fromRational . toRational
 -- Helpers for stateful deserialization
 
 data ParsingLibrary = ParsingLibrary
-    { _images :: Map Text Image
+    { _plImages :: Vector Image
     , _countingIDs :: Map Text (Map Text Int) -- First key is category, second is individual id and associated index
     , _items :: Map Text Item
     , _damageProfiles :: Map Text DamageProfile
@@ -48,7 +51,7 @@ makeLenses ''ParsingLibrary
 
 mkEmptyPL :: SDL.Renderer -> ParsingLibrary
 mkEmptyPL ren = ParsingLibrary
-    { _images = Map.empty
+    { _plImages = Vec.empty
     , _countingIDs = Map.empty
     , _items = Map.empty
     , _damageProfiles = Map.empty
@@ -98,19 +101,25 @@ liftSL = SystemLink . lift
 (.:~?) obj t = Compose (nestedChange <$> obj .:? t)
     where nestedChange = maybe (return Nothing) (\v -> v >>= return . Just)
 
+(.!=~) :: Monad m => Compose Parser m (Maybe a) -> a -> Compose Parser m a
+(.!=~) p def = fmap (fromMaybe def) p
+
 parseUrl :: FilePath -> SystemLink String
 parseUrl path = gets (view workingDirectory) >>= \dir -> pure . resolveAssetPath dir $ path
 
-lookupOrAdd :: Text -> Text -> (Map Text Int -> Map Text Int) -> SystemLink Int
-lookupOrAdd cat key adder = (Map.lookup cat <$> gets (view countingIDs)) >>= checkCat
+lookupOrAddSL :: Text -> Text -> (SystemLink ()) -> SystemLink Int
+lookupOrAddSL cat key adder = (Map.lookup cat <$> gets (view countingIDs)) >>= checkCat
     where checkCat (Just v) = check . Map.lookup key $ v
           checkCat Nothing  = modify (over countingIDs (Map.insert cat Map.empty)) >> check Nothing          
           check (Just v) = pure v
-          check Nothing  = modify addVal >> lookupOrAdd cat key (\_ -> error "Attempting to add value during look up more than once")
-          addVal = over countingIDs (Map.adjust adder cat)
+          check Nothing  = adder >> lookupOrAdd cat key failure
+          failure = pure . error . concat $ ["Attempting to add value during look up more than once (key: ", show key, ", category: ", show cat, ")"]
+
+lookupOrAdd :: Text -> Text -> (Map Text Int -> Map Text Int) -> SystemLink Int
+lookupOrAdd cat key adder = lookupOrAddSL cat key (modify $ over countingIDs (Map.adjust adder cat))
 
 mapInsertIncr :: Ord k => k -> Map k Int -> Map k Int
-mapInsertIncr k m = Map.insert k (Map.size m + 1) m
+mapInsertIncr k m = Map.insert k (Map.size m) m
 
 incrUpdateIDs :: (Ord t, Ord k) => t -> k -> Map t (Map k Int) -> Map t (Map k Int)
 incrUpdateIDs typ k m = Map.adjust updateInner typ newMap