+++ ChangeLog.md
@@ -7,3 +7,10 @@
 ## 0.1.3.0 -- 2019-04-24
 
 * Implemented surface depth sorting
+
+## 0.1.4.0 -- 2019-05-06
+
+* Moved render system to a contained data type
+* Render system can load palettes and switch between them to change the colors of the displayed sprites
+* Additional parsing utilities
+* Damage types can now have defaul resistances for when an actor doesn't specify one